--- src/main/java/com/buri/core/common/Response/Handler/ResponseHandler.java
@@ -1,17 +1,0 @@
-package com.buri.core.common.Response.Handler;
-
-import com.buri.core.common.Response.ResponseDto;
-import com.buri.core.common.Response.ResponseMessage;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ResponseHandler {
-    public <T> ResponseEntity<ResponseDto> toResponseEntity(ResponseMessage message, T data) {
-        return ResponseEntity
-            .status(message.getStatus())
-            .body(
-                new ResponseDto<>(message, data)
-            );
-    }
-}
